Modification of programmatic access to analysis results
Using an interface as introduced in the previous commit is nice, but
does not allow accessing the interface method, if only
AbstractCommitAnalzyer is known. Hence, move that access to that
abstract class and realize a mock-up such that specific analyzer may
override that method, but do not have to implement it.