Modification of (generic) analysis results
In order to be more generic and flexible in defining specific analysis
results for individual analyzers, this commit introduces the
AbstractAnalysisResult class. Specific results may define their
individual sub-classes, like the VerificationRelevantResult. That class
is required for the integration with an incremental KernelHaven variant
and used by the DeadCodeChangeAnalyzer.
Hence, the previously introduced AnalysisResult class is not used
anymore.